refact: YoutubePanel extracted from RoomPage
- Separate Youtube panel extracted from RoomPage for readability
